pyproject.toml
[project]
name = "hack-annotation-guide"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "python-pptx>=1.0.2",
    "httpx>=0.27",
]
